+++ src/pages/Decks/Decks.jsx
@@ -4,14 +4,11 @@ import { MainContent } from './MainContent';
 import { useEffect, useState } from 'react';
 import { supabase } from '../../supabaseClient';
 import { useSearchParams } from 'react-router-dom';
-import { Paginate } from './Paginate';
 
 const Decks = () => {
   const [decks, setDecks] = useState([]);
   const [searchParams, setSearchParams] = useSearchParams();
   const [filters, setFilters] = useState({ qualifications: [], boards: [], subjects: [] });
-  const [currentPage, setCurrentPage] = useState(1);
-  const [postsPerPage] = useState(9);
 
   const filterDecks = (decks) => {
     return decks.filter((deck) => {
@@ -83,26 +80,6 @@ const Decks = () => {
     }));
   };
 
-  const indexOfLastDeck = currentPage * postsPerPage;
-  const indexOfFirstDeck = indexOfLastDeck - postsPerPage;
-  const currentDecks = decks.slice(indexOfFirstDeck, indexOfLastDeck);
-
-  const paginate = (pageNumber) => {
-    setCurrentPage(pageNumber);
-  };
-
-  const previousPage = () => {
-    if (currentPage !== 1) {
-      setCurrentPage(currentPage - 1);
-    }
-  };
-
-  const nextPage = () => {
-    if (currentPage !== Math.ceil(decks.length / postsPerPage)) {
-      setCurrentPage(currentPage + 1);
-    }
-  };
-
   useEffect(() => {
     setSearchParams();
     getBoards();
@@ -118,16 +95,8 @@ const Decks = () => {
       <Container as='section' maxW='8xl' py='50px'>
         <Grid templateColumns='repeat(4, 1fr)'>
           <SidePanel {...filters} />
-          <MainContent currentDecks={currentDecks} />
+          <MainContent decks={decks} />
         </Grid>
-        <Paginate
-          postsPerPage={postsPerPage}
-          totalDecks={decks.length}
-          currentPage={currentPage}
-          paginate={paginate}
-          previousPage={previousPage}
-          nextPage={nextPage}
-        />
       </Container>
     </>
   );
